refactor(authoring): move TreeRowWrapper and the tree model out of VariablesPanel
VariablesPanel 8,472 -> 7,524. A 948-line drop, more than the ~700 projected,
because a whole helper family turned out to be private to `TreeRowWrapper` rather
than shared as it looked.

Against the plan's baseline the three files now stand at:

  VariablesPanel    8,753 -> 7,524  (-1,229)
  InspectorContent  5,547 -> 5,426  (-121)
  InspectorPanel    2,434 -> 2,373  (-61)

This is a **move, not a refactor**. `TreeRowWrapper` turned out to be prop-driven —
no `useEditorStore`, `useBindingAuthoring`, `usePoseRig`, `useAnimationStore` or
`useReferenceFace` — and declared at module scope, so it closed over nothing. It
stays in the feature layer: it discriminates node kinds and casts to
`PoseNodeData`/`RigNodeData`/`PoseGroupNodeData`, which is family C in the plan and
would be rejected from `editor/` by the import boundary anyway.

**Two files, not one, to avoid a runtime cycle.** `variablesTreeModel.ts` holds the
tree-node data model both files need. Importing back from `VariablesPanel` would
have linted clean — `import/no-cycle` is off — while leaving
`VariablesPanel → TreeRowWrapper → VariablesPanel` a real cycle with `const`
scope-class values in TDZ during module init. A third module removes the question.

The surprise: the entire face-ownership family — `createFaceOwnershipSummary`,
`collectNodeFaceOwnership`, `resolveFaceOwnershipScope` and friends — reads as
shared (10 and 9 references) but every reference is *inside the family itself*. Its
only escape into `VariablesPanel` was one call in the row. Hence 948 rather than 700.
The reverse trap was there too: `MAIN_`/`REFERENCE_FACE_SCOPE_ICON_CLASS` look
private to `OwnershipScopeIcon` but are also used by the ownership legend chips, so
the three sibling constants stayed together.

`OwnershipScopeIcon` moved as well — the audit recommended it, and it had exactly
two references.

No behaviour change, verified mechanically rather than by eye: every removed line
was set-diffed against every added line. The only three that differ are
`collectFolderReferenceRigSelectionIds`, `collectFolderReferencePoseSelectionIds`
and `type RigNodeSource`, all because adding `export ` pushed the signature past 80
columns and prettier reflowed a parameter. Every `data-testid`, `aria-label`,
`title` and handler is byte-identical.

Six imports became dead in `VariablesPanel` once the row left and were removed. The
pre-existing `buildCanonicalBindingExpression` warning survives — it was never a
`TreeRowWrapper` dependency.

749 tests, `VariablesPanel.test.tsx` (4,397 lines) passing **unmodified**, typecheck
clean, lint at baseline, prettier clean.
